requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_full_version >= '3.6.0'
django==3.2.5
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1; python_version >= '3.5'
waitress==2.0.0
whitenoise==5.3.0
